# server/docs/swagger.yaml
definitions:
...
  response.Email:
    properties:
      body:
        description: 邮件内容
        type: string
      subject:
        description: 邮件标题
        type: string
      to:
        description: 邮件发送给谁
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  traffic.IspReqParam:
    properties:
      dstProvince:
        type: string
      endTime:
        type: string
      isOversea:
        type: integer
      isp:
        type: string
      ispNameList:
        items:
          type: string
        type: array
      limit:
        description: 每页大小
        type: integer
      linkIdList:
        items:
          type: integer
        type: array
      page:
        description: 页码
        type: integer
      rankLevel:
        description: '@Enum level1,level2,level3'
        type: string
      startTime:
        type: string
      userIdList:
        items:
          type: integer
        type: array
    required:
    - endTime
    - rankLevel
    - startTime
    type: object
...
info:
  contact: {}
paths:
...
  /email/emailTest:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"发送成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 发送测试邮件
      tags:
      - System
  /email/sendEmail:
    post:
      parameters:
      - description: 发送邮件必须的参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/response.Email'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"发送成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 发送邮件
      tags:
      - System
...
  /traffic/isp/export:
    post:
      consumes:
      - application/json
      parameters:
      - description: 开始时间 结束时间 链路
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/traffic.IspReqParam'
      produces:
      - application/json
      responses:
        "200":
          description: excel导出
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 运营商流量导出
      tags:
      - Isp运营商业务流量分析
  /traffic/isp/rankData:
    post:
      consumes:
      - application/json
      parameters:
      - description: 开始时间 结束时间 链路
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/traffic.IspReqParam'
      produces:
      - application/json
      responses:
        "200":
          description: 饼图、趋势图
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 运营商流量查询
      tags:
      - Isp运营商业务流量分析
  /traffic/isp/rankTableData:
    post:
      consumes:
      - application/json
      parameters:
      - description: 开始时间 结束时间 链路
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/traffic.IspReqParam'
      produces:
      - application/json
      responses:
        "200":
          description: 1级 2级 表格
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 运营商流量查询
      tags:
      - Isp运营商业务流量分析
...
swagger: "2.0"